requirements.txt
fastapi
flask
matplotlib
numpy
opencv-python
pillow
python-multipart
tensorflow
uvicorn
transformers
